+++ impl/mutation/mutation.go
@@ -23,7 +23,7 @@ import (
 	cmutation "github.com/google/keytransparency/core/mutation"
 
 	tpb "github.com/google/keytransparency/core/proto/keytransparency_v1_types"
-	spb "github.com/google/keytransparency/impl/proto/mutation_v1_service"
+	spb "github.com/google/keytransparency/core/proto/mutation_v1_service"
 )
 
 // Server holds internal state for the monitor server.
