+++ integration/monitor_test.go
@@ -30,9 +30,9 @@ import (
 	"github.com/google/trillian/crypto/keys/pem"
 	"github.com/google/trillian/crypto/keyspb"
 
+	spb "github.com/google/keytransparency/core/proto/keytransparency_v1_service"
 	kpb "github.com/google/keytransparency/core/proto/keytransparency_v1_types"
-	spb "github.com/google/keytransparency/impl/proto/keytransparency_v1_service"
-	mupb "github.com/google/keytransparency/impl/proto/mutation_v1_service"
+	mupb "github.com/google/keytransparency/core/proto/mutation_v1_service"
 )
 
 const (
